extract webserver & email logic from Envir
